+++ src/urllib3/contrib/webextensions/_async/sse.py
@@ -33,8 +33,13 @@ class AsyncServerSideEventExtensionFromHTTP(AsyncExtensionFromHTTP):
     async def close(self) -> None:
         if self._stream is not None and self._response is not None:
             await self._stream.aclose()
-            if self._response._fp is not None and hasattr(self._response._fp, "abort"):
-                await self._response._fp.abort()
+            if (
+                self._response._fp is not None
+                and self._police_officer is not None
+                and hasattr(self._response._fp, "abort")
+            ):
+                async with self._police_officer.borrow(self._response):
+                    await self._response._fp.abort()
             self._stream = None
             self._response = None
             self._police_officer = None