+++ test/__init__.py
@@ -70,7 +70,7 @@ INVALID_SOURCE_ADDRESSES = [(("192.0.2.255", 0), False), (("2001:db8::1", 0), Tr
 SHORT_TIMEOUT = 0.001
 LONG_TIMEOUT = 0.3
 if os.environ.get("CI") or os.environ.get("GITHUB_ACTIONS") == "true":
-    LONG_TIMEOUT = 0.6
+    LONG_TIMEOUT = 1.5
 
 DUMMY_POOL = ConnectionPool("dummy")